+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Properly parse modular components in Jira queries (OSIDB-4224)
 
+### Changed
+- Swap out local krb5_auth for django-kaminarimon (OSIDB-4243)
+
 ## [4.11.0] - 2025-05-27
 ### Added
 - Add CISA CVSSIssuer type (OSIDB-4203)